correct8

+++ users/models.py
@@ -37,7 +37,11 @@ class Payments(models.Model):
     paid_lesson = models.ForeignKey(
         to=Lesson, on_delete=models.SET_NULL, blank=True, null=True, verbose_name="Оплаченный урок"
     )
-    payment_amount = models.PositiveIntegerField(verbose_name="Сумма оплаты")
+    payment_amount = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        verbose_name="Сумма оплаты"
+    )
     payment_method = models.CharField(max_length=20, verbose_name="Способ оплаты", choices=PAYMENT_METHOD_CHOICES)
     stripe_session_id = models.CharField(max_length=255, blank=True, null=True, verbose_name="ID сессии")
     stripe_payment_intent_id = models.CharField(max_length=255, null=True, blank=True)
